requirements.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
Django==4.2.5
django-click==2.3.0
django-token==0.1.3
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.16.1
rich==13.5.3
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
